Add two scripts used in filtering tidy output.
The first script simply copies all output to a file.  This is a cheap
way of allowing a later buildbot step to process the output of an
earlier step.

The second script parses the output of clang-tidy, looking for
warnings caused by Qt header files.  These warnings aren't fixable by
us, so strip them from the the output.  This will leave a set of
warnings that we can actually fix.